Bootstrap base SaaS library
Summary:
Those base classes have been extracted from the MediaWiki SaaS
entry point.

The goal of the service is to check if the instance exists,
and then to load the instance required and specific configuration.

It also features a /status entry point to allow easy querying
from monitoring services.

The class methods have been inspired by the API microservice,
so a common foundation for a microservice and a SaaS could be
possible if we need to make them evolve more in the future.

Ref T845.

Test Plan:
Those classes are used by saas-mediawiki, so they work with
a real implementation.

Unit tests cover helper methods.

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T845

Differential Revision: https://devcentral.nasqueron.org/D1443